+++ CHANGELOG.md
@@ -5,8 +5,9 @@ http://phpsysinfo.sourceforge.net/
 
 phpSysInfo 3.1.16
 ----------------
- - [ADD] Show number of process for Linux, FreeBSD, Darwin, Windows, Minix and Android
 
+ - [ADD] Show number of process for Linux, FreeBSD, Darwin, Windows, Minix and Android
+ - [ADD] 4MLinux to detected distros
 
 phpSysInfo 3.1.15
 ----------------