+++ CHANGELOG.md
@@ -6,7 +6,7 @@ http://phpsysinfo.sourceforge.net/
 phpSysInfo 3.1.16
 ----------------
 
- - [ADD] Show number of process for Linux, FreeBSD, NetBSD, OpenBSD, DragonFly, Darwin, Windows, Minix and Android
+ - [ADD] Show number of process
  - [ADD] 4MLinux to detected distros
 
 phpSysInfo 3.1.15