+++ CHANGELOG.md
@@ -3,6 +3,12 @@ Changelog of phpSysInfo
 
 http://phpsysinfo.sourceforge.net/
 
+phpSysInfo 3.1.16
+----------------
+
+ - [ADD] Show number of process
+ - [ADD] Lunar and 4MLinux to detected distros
+
 phpSysInfo 3.1.15
 ----------------